% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual109401.R
\docType{data}
\name{actual109401}
\alias{actual109401}
\title{Dataset required by medulloClassify}
\format{An object of class \code{character} of length 19.}
\usage{
actual109401
}
\description{
This dataset contains an expression matrix where the rownames are HUGO gene
symbols and the column names contain the sample identifiers. Dataset also contains a
character vector containing the observations (i.e. actual classifications) made for each sample.
}
\keyword{datasets}
