% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_109401.R
\docType{data}
\name{actual_109401}
\alias{actual_109401}
\title{Example Dataset for testing medulloClassify}
\format{
An object of class \code{character} of length 19.
}
\usage{
actual_109401
}
\description{
Dataset is a character vector containing the observations (i.e. actual classifications)
made for each sample. The vector can contain the following characters: "WNT", "SHH", "Group3",
"Group4", and "U".
}
\keyword{datasets}
