% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestFeaturesNew.R
\docType{data}
\name{bestFeaturesNew}
\alias{bestFeaturesNew}
\title{Dataset required by medulloClassify}
\format{A character vector of 279 genes}
\usage{
bestFeaturesNew
}
\description{
This dataset contains a vector of 279 genes
}
\keyword{datasets}
