% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScore.R
\name{calcScoreMat}
\alias{calcScoreMat}
\title{Calc score}
\usage{
calcScoreMat(myMat = NULL, mySetsUp = NULL)
}
\arguments{
\item{myMat}{matrix containing gene ratios for each gene signature and sample.}

\item{mySetsUp}{list of 4 containing the gene ratios associated with
each of the 4 molecular subtypes of Medulloblastoma.}
}
\description{
Function to calculate column mean for each sample in myMat Matrix.
}
\details{
None
}
\author{
Pichai Raman
}
