% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify}
\usage{
classify(geneRatioOut = NULL,
  medulloGeneSetsUp = "../../data/medulloSetsUp.RDS")
}
\arguments{
\item{geneRatioOut}{Matrix containing the gene ratios for different samples.}

\item{medulloGeneSetsUp}{list of 4 containing the gene ratios associated with
each of the 4 molecular subtypes of Medulloblastoma.}
}
\description{
Function to classify
}
\details{
Classifier for predicting amongst the 4 molecular subtypes of
Medulloblastoma, Sonic Hedgehog (SHH), WNT, Group 3, and Group 4 from
RNA-Seq or microarray data.
}
\author{
Pichai Raman

Sherjeel Arif
}
