% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneRatioOut_109401.R
\docType{data}
\name{geneRatioOut_109401}
\alias{geneRatioOut_109401}
\title{Example Dataset for testing calcScore() function}
\format{An object of class \code{data.frame} with 772 rows and 19 columns.}
\usage{
geneRatioOut_109401
}
\description{
Dataset is a data frame containing expression values of the gene ratios with respect to each
sample. The rownames are the gene ratio symbols and the column names are the sample identifiers.
}
\keyword{datasets}
