% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medulloSetsUp.R
\docType{data}
\name{medulloSetsUp}
\alias{medulloSetsUp}
\title{Dataset required by medulloClassify}
\format{
A list of 4 Medulloblastoma subtypes with gene signatures
\describe{
  \item{WNT}{character vector of 60 genes}
  \item{SHH}{character vector of 20 genes}
  \item{Group3}{character vector of 158 genes}
  \item{Group4}{character vector of 274 genes}
}
}
\usage{
medulloSetsUp
}
\description{
This dataset contains a list of 4 Medulloblastoma subtypes with gene signatures
}
\keyword{datasets}
