% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureGenes.R
\name{signatureGenes}
\alias{signatureGenes}
\title{Signature Genes}
\arguments{
\item{exprs}{expression matrix in which the rows contains the gene symbols and the
columns are the samples. The values inside the matrix are the respective gene signatures.}
}
\description{
Function to calculate gene ratio matrix
}
\details{
Function for determining the gene ratio matrix. The function expects an expression matrix
as the only parameter.
}
\author{
Pichai Raman
}
