% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStats.R
\name{calcStats}
\alias{calcStats}
\title{Calc Stats}
\usage{
calcStats(myClassActual = NULL, myClassPred = NULL)
}
\arguments{
\item{myClassActual}{a vector of characters corresponding to the actual
classification for each sample in the dataset.}

\item{myClassPred}{a vector of characters corresponding to the prediction made
for each sample.}
}
\description{
Function to calculate statistics on prediction class and
actual class of a medulloblastoma dataset.
}
\details{
This function can be used to analyze prediction class if the correct
classification of the provided Medulloblastoma dataset is known. Function
calculates an accuracy score, a confidence interval, creates a confusion
matrix, and displays other useful statistics (i.e. Sensitivity, Specificity).

The parameters are character vectors and each value in the vector corresponds to
one of the four molecular subtypes of Medulloblastoma (i.e. Sonic Hedgehog (SHH),
WNT, Group 3, and Group 4).
}
\examples{
## mock prediction class
predictionClass <- c("Group3", "WNT", "WNT", "Group4", "SHH", "Group4")

## mock actual class
actualClass <- c("Group4", "WNT", "SHH", "Group4", "SHH", "Group4")

## call function to obtain accuracy and other statistics
medulloPackage::calcStats(actualClass, predictionClass)

}
\author{
Pichai Raman

Sherjeel Arif
}
