% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify}
\arguments{
\item{geneRatioOut}{matrix containing gene ratios for each gene signature and sample.
The row names contain the gene signatures and the column names contain the individual samples.}

\item{medulloGeneSetsUp}{list of 4 containing the gene signature associated with
each of the 4 molecular subtypes of Medulloblastoma.}
}
\description{
Function to classify
}
\details{
Classifier for predicting amongst the 4 molecular subtypes of
Medulloblastoma, Sonic Hedgehog (SHH), WNT, Group 3, and Group 4 from
RNA-Seq or microarray data.
}
\author{
Pichai Raman

Sherjeel Arif
}
