% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify}
\usage{
classify(exprs = NULL, medulloGeneSetsUp = NULL)
}
\arguments{
\item{exprs}{matrix containing gene expression values.
The row names contain HUGO gene symbols and the column names contain the sample identifiers.}

\item{medulloGeneSetsUp}{list of 4 containing the gene signature associated with
each of the 4 molecular subtypes of Medulloblastoma.}
}
\description{
Function to classify
}
\details{
Classifier for predicting amongst the 4 molecular subtypes of
Medulloblastoma, Sonic Hedgehog (SHH), WNT, Group 3, and Group 4 from
RNA-Seq or microarray data.
}
\author{
Pichai Raman

Sherjeel Arif

Komal Rathi
}
