% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs_109401.R
\docType{data}
\name{exprs_109401}
\alias{exprs_109401}
\title{Example Dataset for testing medulloClassify}
\format{An object of class \code{data.frame} with 12233 rows and 19 columns.}
\usage{
exprs_109401
}
\description{
Dataset contains an expression matrix where the rownames are HUGO gene
symbols and the column names contain the sample identifiers.
}
\keyword{datasets}
